pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "crs"
version = "1.0.0"
requires-python = ">=3.12,<4.0"
description = "Theori's CRS"

dependencies = [
    "aiofile==3.9.0",
    "aiosqlite==0.21.0",
    "cxxfilt==0.3.0",
    "fastapi==0.115.8",
    "flake8==7.2.0",
    "flake8-async==25.3.1",
    "google-auth==2.39.0",
    "jsonpickle==3.3.0",
    "influxdb-client==1.48.0",
    "litellm[proxy]==1.67.2",
    "lxml==5.3.0",
    "mypy==1.11.2",
    "numpy==2.1.2",
    "numpydoc==1.8.0",
    "opentelemetry-api==1.30.0",
    "opentelemetry-exporter-otlp==1.30.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.30.0",
    "opentelemetry-sdk==1.30.0",
    "pydantic==2.10.3",
    "pygit2==1.17.0",
    "pyright==1.1.402",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "python-magic==0.4.27",
    "PyYAML==6.0.2",
    "quicksect==0.2.2",
    "requests==2.32.3",
    "result==0.17.0",
    "rich==13.9.4",
    "ruff==0.11.2",
    "rustworkx==0.16.0",
    "types-requests==2.32.0.20241016",
    "types-PyYAML==6.0.12.20240917",
    "types-lxml==2024.11.8",
    "tenacity==9.0.0",
    "tqdm==4.67.1",
    "tree-sitter==0.24.0",
    "tree-sitter-c==0.23.4",
    "tree-sitter-cpp==0.23.4",
    "tree-sitter-java==0.23.5",
    "uvicorn==0.29.0",
    "orjson>=3.10.18",
    "py-spy>=0.4.0",
    "viztracer>=1.0.4",
    "maturin>=1.8.6",
    "jinja2>=3.1.6",
]

[tool.setuptools]
py-modules = ["crs"]

[tool.mypy]
strict = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
enable_incomplete_feature = "NewGenericSyntax"
disable_error_code = "unused-ignore"

[tool.pyright]
include = ["crs", "eval.py", "create_tasks.py", "round_sim.py"]
typeCheckingMode = "strict"  # Options: "off", "basic", "strict"
reportUnusedCallResult = "error"

[tool.ruff]
include = ["crs/**/*.py", "crs/*.py"]
target-version = "py312"

[tool.ruff.lint]
ignore = ["F405", "F402", "F403", "E402", "F541", "E701", "E741", "RUF100"]

# flake8 config not possible in pyproject...

[tool.pytest.ini_options]
# Specify the directory where your tests are located
testpaths = [ "tests" ]
filterwarnings = ["ignore::DeprecationWarning", "ignore::pytest.PytestCollectionWarning"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
markers = [ "slow: marks tests as slow (deselect with '-m \"not slow\"')" ]

[tool.maturin]
features = ["pyo3/extension-module"]
module-name = "crs_rust"
include = ["src/crs_rust.pyi"]
